improving change callbacks
Some frameworks - like 'chosen' - creates visual elements to represent real selects.
When an option is marked as 'selected' programmatically, it don't fire 'change' event.
This fix trigger 'change' event whenever is necessaray, so it can be listened by others.
It was necessary also to create a 'onAfterReset' to allow others to intercept the first
moment a select turns 'disabled' as triggering 'change' without changing the selected
option doesn't cause event 'change' to be triggered properly (because nothing changes
of course).